A lot of progress was done on the NeweggParser. It now reads the categories and the subcategories. Next step is to add the ability to make HTTP POST requests so we can get the list of products for each category.
A lot of restructuring was also done. The core is now a separate project
from the modules, and so you will find Eclipse project files in the
backend folder, as well as in the modules/newegg folder (and so you have
to create two Eclipse projects). You should probably delete your old
projects and re-add them.