Added basic MariaDB driver implementation.
There is a new database/ folder in the core that contains the SQL script
to create the appropriate database and schemas as well as a
configuration script with the parameters needed to connect to it. When
the application is deployed, the init script needs to be run under root
MariaDB account and the permissions need to be granted to the
appropriate user. This is tested and works on Ubuntu 12.10.

The Makefile for the Newegg module has also been updated to reflect the
JSoup library (this seems to have been forgotten to be added).