Fixed the bug with InterruptableInputStream so we don't have to arbitrarily sleep in the module (and wait for the Core to recognize the existence of our response). StreamPipe now uses InterruptableInputStream.
Makefile is changed, but still does not build the core oddly. Needs
fixing. The MariaDBDriver class needs implementing. The next logical
change to make for the core is to have it read module configuration from
a settings file and run them on separate threads. These are launched
every now and then with either a product list query or a product info
query.